requirements.txt
Django==1.11
django-cors-headers==3.2.1
djangorestframework==3.11.0